+++ docassemble_base/docassemble/base/astparser.py
@@ -43,6 +43,8 @@ class myvisitnode(ast.NodeVisitor):
         self.depth -= 1
     def visit_Call(self, node):
         self.calls.add(node.func)
+        if hasattr(node.func, 'id') and node.func.id in ['showif', 'showifdef', 'value', 'defined'] and len(node.args) and node.args[0].__class__.__name__ == 'Str' and hasattr(node.args[0], 's'):
+            self.names[node.args[0].s] = 1
         ast.NodeVisitor.generic_visit(self, node)
     def visit_Attribute(self, node):
         if node not in self.calls: